requirements.txt
https://github.com/Relanit/TwitchIO/archive/5c0524a.zip
python-dateutil~=2.8.2
dnspython==2.2.0
pycountry~=22.3.5
aiohttp~=3.7.4.post0
motor~=3.0.0
beautifulsoup4~=4.11.1